Add: MyPageSide, MyPageHeader Component 생성
