+++ src/Pages/MyPage/Components/Aside/MyPageAside.scss
@@ -1,0 +1,40 @@
+@import '../../../../Styles/common.scss';
+
+.MyPageAside {
+  letter-spacing: 0.03em;
+
+  .title {
+    display: inline-block;
+    width: 160px;
+    margin-bottom: 35px;
+    padding: 0 0 15px 15px;
+    font-size: 17px;
+    font-weight: 700;
+    border-bottom: solid 1px rgba(211, 211, 211, 0.5);
+  }
+
+  .subTitle {
+    margin-bottom: 35px;
+    font-weight: 700;
+    font-size: 13px;
+
+    .subCategory {
+      //   color: gray;
+      padding: 9px;
+      padding-left: 15px;
+      font-weight: 400;
+      font-size: 11px;
+    }
+
+    .subCategory:hover {
+      font-weight: 700;
+      //   background-color: black;
+      background-color: rgb(248, 248, 248);
+      cursor: pointer;
+    }
+  }
+
+  .subTitle .subCategory:first-child {
+    margin-top: 9px;
+  }
+}
